requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
dotenv==0.9.9
pytest==8.3.3
pytest-django==4.9.0
mkdocs==1.6.1